Conflicts:
	.gitignore
	build.xml
	nbproject/build-impl.xml
	nbproject/genfiles.properties
	nbproject/project.properties
	nbproject/project.xml
	src/edu/esprit/gui/GestionChauffeur.form
	src/edu/esprit/gui/GestionChauffeur.java

+++ src/edu/esprit/DAO/classes/ReclamationDAO.java
@@ -37,66 +37,119 @@ public class ReclamationDAO implements IReclamationDAO{
     }
     
     @Override
-    public void insertReclamation(Reclamation reclamation) {
-        String requete = "insert into reclamation values (?,?,?,?,?)";
+    public boolean isReclamationExist(Reclamation r) {
+    String requete = "select * from reclamation where idReclamation=?";
+
         try {
             PreparedStatement ps = conn.prepareStatement(requete);
-            ps.setString(1, reclamation.getIdReclamation());
-            ps.setString(2, reclamation.getClient().getIdClient());
-            ps.setString(3, reclamation.getAgence().getIdAgence());
+ 
+            ps.setInt(1, r.getIdReclamation());
+            
+            ResultSet resultat = ps.executeQuery();
+            return resultat.next();
+        }
+        catch (SQLException ex) {
+            System.out.println("erreur lors du chargement du reclamation" + ex.getMessage());
+            return false;
+        }
+                
+    }
+    
+    @Override
+    public Boolean insertReclamation(Reclamation reclamation) {
+        if(isReclamationExist(reclamation))
+            return false;
+        String requete = "insert into reclamation values (?,?,?,?,?,?)";
+        try {
+            PreparedStatement ps = conn.prepareStatement(requete);
+            
+            ps.setInt(1, reclamation.getIdReclamation());
+            ps.setInt(2, reclamation.getClient().getIdClient());
+            ps.setInt(3, reclamation.getAgence().getIdAgence());
             ps.setString(4, reclamation.getsujetReclamation());
             ps.setString(5, reclamation.getcontenu());
+            ps.setBoolean(6, reclamation.isEtat());
 
             ps.executeUpdate();
             System.out.println("Ajout effectuée avec succès");
+            return true;
         } catch (SQLException ex) {
             //Logger.getLogger(PersonneDao.class.getName()).log(Level.SEVERE, null, ex);
             System.out.println("erreur lors de l'insertion dans reclamation " + ex.getMessage());
+            return false;
         }
     }
 
     @Override
-    public void updateReclamation(Reclamation reclamation) {
-        String requete = "update reclamation set idClient=?, idAgence=?, sujetReclamation=?, contenu=? where idReclamation=?";
+    public Boolean updateReclamation(Reclamation reclamation) {
+        if(!isReclamationExist(reclamation))
+            return false;
+        String requete = "update reclamation set sujetReclamation=?, contenu=?, etat=? where idReclamation=?";
         try {
             PreparedStatement ps = conn.prepareStatement(requete);
             
-            ps.setString(5, reclamation.getIdReclamation());
-            ps.setString(1, reclamation.getClient().getIdClient());
-            ps.setString(2, reclamation.getAgence().getIdAgence());
-            ps.setString(3, reclamation.getsujetReclamation());
-            ps.setString(4, reclamation.getcontenu());
-            
+            ps.setInt(4, reclamation.getIdReclamation());
+            ps.setString(1, reclamation.getsujetReclamation());
+            ps.setString(2, reclamation.getcontenu());
+            ps.setBoolean(3, reclamation.isEtat());
              
             ps.executeUpdate();
             System.out.println("Mise à jour effectuée avec succès");
+            return true;
         } catch (SQLException ex) {
             //Logger.getLogger(PersonneDao.class.getName()).log(Level.SEVERE, null, ex);
             System.out.println("erreur lors de la mise à jour de réclamation " + ex.getMessage());
+            return false;
         }
-
+    }
+    
+    @Override
+    public Boolean updateReclamationRespAgence(Reclamation reclamation) {
+        if(!isReclamationExist(reclamation))
+            return false;
+        String requete = "update reclamation set etat=? where idReclamation=?";
+        try {
+            PreparedStatement ps = conn.prepareStatement(requete);
+            
+            ps.setInt(2, reclamation.getIdReclamation());
+            ps.setBoolean(1, reclamation.isEtat());
+             
+            ps.executeUpdate();
+            System.out.println("Mise à jour effectuée avec succès");
+            return true;
+        } catch (SQLException ex) {
+            //Logger.getLogger(PersonneDao.class.getName()).log(Level.SEVERE, null, ex);
+            System.out.println("erreur lors de la mise à jour de réclamation " + ex.getMessage());
+            return false;
+        }
     }
 
     @Override
-    public void deleteReclamation(String id) {
+    public Boolean deleteReclamation(Reclamation reclamation) {
+        if(isReclamationExist(reclamation))
+            return false;
         String requete = "delete from reclamation where idReclamation=?";
         try {
             PreparedStatement ps = conn.prepareStatement(requete);
-            ps.setString(1, id);
+            ps.setInt(1, reclamation.getIdReclamation());
             ps.executeUpdate();
             System.out.println("Suppression effectuée avec succès");
+            return true;
         } catch (SQLException ex) {
             //Logger.getLogger(PersonneDao.class.getName()).log(Level.SEVERE, null, ex);
             System.out.println("erreur lors de la suppression du reclamation" + ex.getMessage());
+            return false;
         }
     }
 
     @Override
-    public Reclamation findReclamationById(String id) {
-        String requete = "select * from trajet where idReclamation=?";
-
+    public Reclamation findReclamationById(int id) {
+        String requete = "select * from reclamation where idReclamation=?";
+        boolean b=false;
         try {
+            b=true;
             PreparedStatement ps = conn.prepareStatement(requete);
+            ps.setInt(1,id);
             ResultSet resultat = ps.executeQuery();
             Reclamation reclamation = new Reclamation();
             ClientDAO clientDAO = ClientDAO.getInstance();
@@ -104,11 +157,13 @@ public class ReclamationDAO implements IReclamationDAO{
             while (resultat.next()) {
                 
                 reclamation.setIdReclamation(id);
-                reclamation.setClient(clientDAO.findClientById(resultat.getString("idClient")));
-                reclamation.setAgence(agenceDAO.findAgenceById(resultat.getString("idAgence")));
+                reclamation.setClient(clientDAO.findClientById(resultat.getInt("idClient")));
+                reclamation.setAgence(agenceDAO.findAgenceById(resultat.getInt("idAgence")));
                 reclamation.setsujetReclamation(resultat.getString("sujetReclamation"));
                 reclamation.setcontenu(resultat.getString("contenu"));
             }
+            if(b==false)
+                return null;
             return reclamation;
         } catch (SQLException ex) {
             System.out.println("erreur lors du chargement du reclamation" + ex.getMessage());
@@ -119,9 +174,10 @@ public class ReclamationDAO implements IReclamationDAO{
     @Override
     public List<Reclamation> DisplayAllReclamations() {
         List<Reclamation> listeReclamations = new ArrayList<>();
-
+        boolean b=false;
         String requete = "select * from reclamation";
         try {
+            b=true;
             Statement statement = conn.createStatement();
             ResultSet resultat = statement.executeQuery(requete);
             Reclamation reclamation = new Reclamation();
@@ -129,14 +185,17 @@ public class ReclamationDAO implements IReclamationDAO{
             AgenceDAO agenceDAO = AgenceDAO.getInstance();
             while (resultat.next()) {
                 
-                reclamation.setIdReclamation(resultat.getString("idTrajet"));
-                reclamation.setClient(clientDAO.findClientById(resultat.getString("idClient")));
-                reclamation.setAgence(agenceDAO.findAgenceById(resultat.getString("idAgence")));
+                reclamation.setIdReclamation(resultat.getInt("idReclamation"));
+                reclamation.setClient(clientDAO.findClientById(resultat.getInt("idClient")));
+                reclamation.setAgence(agenceDAO.findAgenceById(resultat.getInt("idAgence")));
                 reclamation.setsujetReclamation(resultat.getString("sujetReclamation"));
                 reclamation.setcontenu(resultat.getString("contenu"));
+                reclamation.setEtat(resultat.getBoolean("etat"));
                
                 listeReclamations.add(reclamation);
             }
+            if(b==false)
+                return null;
             return listeReclamations;
         } catch (SQLException ex) {
             //Logger.getLogger(PersonneDao.class.getName()).log(Level.SEVERE, null, ex);
